pyproject.toml
[project]
name = "routine-grid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "psycopg[binary]>=3.2.7",
    "python-dotenv>=1.1.0",
]
